+++ Blocks/Air.cs
@@ -7,9 +7,9 @@ namespace Blocks
 {
     public sealed class Air : Block
     {
-        public override ulong MaxStateId => 0;
+        public override int MaxStateId => 0;
 
-        public override ulong MinStateId => 0;
+        public override int MinStateId => 0;
 
         public override Block Clone()
         {
@@ -20,6 +20,6 @@ namespace Blocks
             };
         }
 
-        public override ulong GetStateId() => 0;
+        public override int GetStateId() => 0;
     }
 }
